requirements.txt
Flask==2.2.3
Flask-SocketIO==5.3.3
gunicorn==20.1.0
eventlet==0.30.2
python-socketio[asyncio_client]==5.8.0
redis==4.5.4